Conduct Randomization Inference

This function makes it easy to conduct three kinds of randomization inference.

conduct_ri(formula = NULL, model_1 = NULL, model_2 = NULL,
  test_function = NULL, assignment = "Z", outcome = NULL, declaration,
  sharp_hypothesis = 0, IPW = TRUE, IPW_weights = NULL,
  sampling_weights = NULL, permutation_matrix = NULL, data, sims = 1000)

Arguments

formula
an object of class formula, as in lm. Use formula when conducting significance tests of an Average Treatment Effect estimate under a sharp null hypothesis. For the difference-in-means estimate, do not include covariates. For the OLS covariate-adjusted estimate, include covariates.

model_1
an object of class formula, as in lm. Models 1 and 2 must be "nested." model_1 should be the "restricted" model and model_2 should be the "unrestricted" model.

model_2
an object of class formula, as in lm. Models 1 and 2 must be "nested." model_1 should be the "restricted" model and model_2 should be the "unrestricted" model.

test_function
A function that takes data and returns a scalar test statistic.

assignment
a character string that indicates which variable is randomly assigned. Defaults to "Z".

outcome
a character string that indicates which variable is the outcome variable. Defaults to NULL.

declaration
A random assignment declaration, created by declare_ra.

sharp_hypothesis
either a numeric scalar or a numeric vector of length k - 1, where k is the number of treatment conditions. In a two-arm trial, this number is the *hypothesized* difference between the treated and untreated potential potential outcomes for each unit.. In a multi-arm trial, each number in the vector is the hypothesized difference in potential outcomes between the baseline condition and each successive treatment condition.

IPW
logical, defaults to TRUE. Should inverse probability weights be calculated?

IPW_weights
a character string that indicates which variable is the existing inverse probability weights vector. Usually unnecessary, as IPW weights will be incorporated automatically if IPW = TRUE. Defaults to NULL.

sampling_weights
a character string that indicates which variable is the sampling weights vector. Optional, defaults to NULL. NOT YET IMPLEMENTED

permutation_matrix
An optional matrix of random assignmnets, typically created by obtain_permutation_matrix.

data
A data.frame.

sims
the number of simulations. Defaults to 1000.

Details

1. Conduct hypothesis tests under the sharp null when the test statistic is the difference-in-means or covariate-adjusted average treatment effect estimate. 2. Conduct "ANOVA" style hypothesis tests, where the f-statistic from two nested models is the test statistic. This procedure is especially helpful when testing interaction terms under null of constant effects. 3. Arbitrary (scalar) test statistics